requirements.txt
django
django-guardian
django-bootstrap4
django-sql-explorer
django-allauth
django-crispy-forms
django-allauth